backend/requirements/development.txt
-r base.txt

# Testing
pytest==8.*
pytest-django==4.9.*
pytest-cov==5.*
pytest-factoryboy==2.*
factory-boy==3.3.*
faker==30.*

# Debug
django-debug-toolbar==4.*
django-extensions==3.*
ipython==8.*

# Code Quality
ruff==0.9.*
mypy==1.14.*
django-stubs==5.*
djangorestframework-stubs==3.*

# Dev tools
watchfiles==1.*
